pyproject.toml
[tool.poetry]
name = "profile-intelligence-engine"
version = "0.1.0"
description = "Profile Intelligence Engine (MVP)"
authors = ["Arjun Kshirsagar <kshirsagararjun20@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.0"
uvicorn = {extras = ["standard"], version = "0.30.6"}
httpx = ">=0.28.1,<1.0"
sqlalchemy = "2.0.35"
psycopg = {extras = ["binary"], version = ">=3.2.1"}
pydantic = "^2.12.0"
pydantic-settings = "2.5.2"
openai = "1.47.0"
google-genai = ">=1.5.0,<2.0.0"
alembic = "^1.18.4"
anthropic = "^0.84.0"
celery = "^5.6.2"
redis = "^7.2.1"
flower = "^2.0.1"
ddgs = "^9.11.2"
groq = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "python-json-logger (>=4.0.0,<5.0.0)"
]
